radius and speed time updates
speed commands were returning an error from the robot because of
incorrect 'min_time' key.

move and other commands were not accepting radius info. It is not
included in the library and sent to the robot, but I'm not sure the
robot pays attention to radius info in commands sent over the network.